% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_filters.R
\name{colour_filters}
\alias{colour_filters}
\title{Apply colour filter to given image}
\usage{
colour_filters(image, tone = "grayscale")
}
\arguments{
\item{image}{string: The local file path for image to which filter will be applied}

\item{tone}{string: Colour filter to be applied to the image. Options:'grayscale', 'blue_tone', 'green_tone', 'red_tone', 'negative'. Default: 'grayscale'}
}
\value{
array: altered image array returned for the input colour filter
}
\description{
Saves the given image in the current working directory with the
user-specified color filter applied and returns the altered image array.
}
