% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_detection.R
\name{edge_detection}
\alias{edge_detection}
\title{Apply edge detection to given image}
\usage{
edge_detection(image_path)
}
\arguments{
\item{image_path}{string: The local file path for an image of png, jpg, or jpeg format to which edge detection filter will be applied.}
}
\value{
array: image array with edge_detection filter applied.
}
\description{
Returns the given image with greyscale edge detection applied.
}
